move formatTime to util
